move index related methods to one class
